+++ pkgs/sdk/server-ai/src/Config/LdAiConfig.cs
@@ -45,17 +45,19 @@ namespace LaunchDarkly.Sdk.Server.Ai.Config
             private readonly List<Message> _prompt;
             private bool _enabled;
             private readonly Dictionary<string, object> _modelParams;
+            private readonly string _key;
 
 
             /// <summary>
             /// Constructs a new builder. By default, the config will be disabled, with no prompt
             /// messages or model parameters.
             /// </summary>
-            public Builder()
+            public Builder(string key)
             {
                 _enabled = false;
                 _prompt = new List<Message>();
                 _modelParams = new Dictionary<string, object>();
+                _key = key;
             }
 
             /// <summary>
@@ -111,7 +113,7 @@ namespace LaunchDarkly.Sdk.Server.Ai.Config
             /// <returns>a new LdAiConfig</returns>
             public LdAiConfig Build()
             {
-                return new LdAiConfig(_enabled, _prompt, new Meta(), _modelParams);
+                return new LdAiConfig(_enabled, _key, _prompt, new Meta(), _modelParams);
             }
         }
 
@@ -125,12 +127,15 @@ namespace LaunchDarkly.Sdk.Server.Ai.Config
         /// </summary>
         public readonly IReadOnlyDictionary<string, object> Model;
 
-        internal LdAiConfig(bool enabled, IEnumerable<Message> prompt, Meta meta, IReadOnlyDictionary<string, object> model)
+
+
+        internal LdAiConfig(bool enabled, string key, IEnumerable<Message> prompt, Meta meta, IReadOnlyDictionary<string, object> model)
         {
             Model = model ?? new Dictionary<string, object>();
             Prompt = prompt?.ToList() ?? new List<Message>();
             VersionKey = meta?.VersionKey ?? "";
             Enabled = enabled;
+            Key = key;
         }
 
 
@@ -153,9 +158,16 @@ namespace LaunchDarkly.Sdk.Server.Ai.Config
         public string VersionKey { get; }
 
         /// <summary>
+        /// The key of this config.
+        /// </summary>
+        public string Key { get; }
+
+
+        /// <summary>
         /// Convenient helper that returns a disabled LdAiConfig.
         /// </summary>
         public static LdAiConfig Disabled = New().Disable().Build();
+
 
     }
 }
