+++ pkgs/sdk/server-ai/src/LdAiClient.cs
@@ -5,199 +5,168 @@ using System.Linq;
 using System.Text.Json;
 using LaunchDarkly.Sdk.Server.Ai.Config;
 using LaunchDarkly.Sdk.Server.Ai.DataModel;
+using LaunchDarkly.Sdk.Server.Ai.Interfaces;
 using Mustache;
 
-namespace LaunchDarkly.Sdk.Server.Ai
+namespace LaunchDarkly.Sdk.Server.Ai;
+
+/// <summary>
+/// The LaunchDarkly AI client. The client is capable of retrieving AI configurations from LaunchDarkly,
+/// and generating events specific to usage of the AI configuration when interacting with model providers.
+/// </summary>
+public sealed class LdAiClient : IDisposable
 {
+    private readonly ILaunchDarklyClient _client;
+
     /// <summary>
-    /// TBD
+    /// Constructs a new LaunchDarkly AI client.
+    ///
+    /// Example:
+    /// <code>
+    /// var config = Configuration.Builder("my-sdk-key").Build();
+    /// var client = new LdClient(config);
+    /// var aiClient = new LdAiClient(client);
+    /// </code>
+    ///
     /// </summary>
-    public interface ILogger
+    /// <param name="client">a LaunchDarkly Server-side SDK client instance</param>
+    public LdAiClient(ILaunchDarklyClient client)
     {
-
-        /// <summary>
-        /// TBD
-        /// </summary>
-        /// <param name="format"></param>
-        /// <param name="allParams"></param>
-        void Error(string format, params object[] allParams);
-
-        /// <summary>
-        /// TBD
-        /// </summary>
-        /// <param name="format"></param>
-        /// <param name="allParams"></param>
-        void Warn(string format, params object[] allParams);
+        _client = client ?? throw new ArgumentNullException(nameof(client));
     }
 
-    /// <summary>
-    /// TBD
-    /// </summary>
-    public interface ILaunchDarklyClient
-    {
-        /// <summary>
-        /// TBD
-        /// </summary>
-        /// <param name="key"></param>
-        /// <param name="context"></param>
-        /// <param name="defaultValue"></param>
-        /// <returns></returns>
-        EvaluationDetail<LdValue> JsonVariationDetail(string key, Context context, LdValue defaultValue);
-
-
-        /// <summary>
-        /// TBD
-        /// </summary>
-        /// <param name="name"></param>
-        /// <param name="context"></param>
-        /// <param name="data"></param>
-        /// <param name="metricValue"></param>
-        void Track(string name, Context context, LdValue data, double metricValue);
-
-        /// <summary>
-        /// TBD
-        /// </summary>
-        void Dispose();
-
-        /// <summary>
-        /// TBD
-        /// </summary>
-        /// <returns></returns>
-        ILogger GetLogger();
-
-    }
+    // This is the special Mustache variable that can be used in prompts to access the current
+    // LaunchDarkly context. For example, {{ ldctx.key }} will return the context key.
+    private const string LdContextVariable = "ldctx";
 
     /// <summary>
-    /// TBD
+    /// Retrieves a LaunchDarkly AI config identified by the given key. The return value
+    /// is an <see cref="LdAiConfigTracker"/>, which makes the configuration available and
+    /// provides convenience methods for generating events related to model usage.
+    ///
+    /// Any variables provided will be interpolated into the prompt's messages.
+    /// Additionally, the current LaunchDarkly context will be available as 'ldctx' within
+    /// a prompt message.
+    ///
     /// </summary>
-    public sealed class LdAiClient : IDisposable
+    /// <param name="key">the flag key</param>
+    /// <param name="context">the context</param>
+    /// <param name="defaultValue">the default config, if unable to retrieve from LaunchDarkly</param>
+    /// <param name="variables">the list of variables used when interpolating the prompt</param>
+    /// <returns>an AI config tracker</returns>
+    public LdAiConfigTracker ModelConfig(string key, Context context, LdAiConfig defaultValue,
+        IReadOnlyDictionary<string, object> variables = null)
     {
-        private readonly ILaunchDarklyClient _client;
 
-        /// <summary>
-        /// TBD
-        /// </summary>
-        /// <param name="client">an ILaunchDarklyClient</param>
-        public LdAiClient(ILaunchDarklyClient client)
+        var detail = _client.JsonVariationDetail(key, context, LdValue.Null);
+
+        var defaultTracker = new LdAiConfigTracker(_client, defaultValue, context, key);
+
+        if (detail.IsDefaultValue)
         {
-            _client = client ?? throw new ArgumentNullException(nameof(client));
-        }
-
-        private static string InterpolateTemplate(string template, IReadOnlyDictionary<string, object> variables)
-        {
-            return Template.Compile(template).Render(variables);
-        }
-
-        private static IDictionary<string, object> GetAllAttributes(Context context)
-        {
-            var attributes = new Dictionary<string, object>();
-            foreach (var key in context.OptionalAttributeNames)
-            {
-                attributes[key] = ValueToObject(context.GetValue(AttributeRef.FromLiteral(key)));
-            }
-
-            attributes["kind"] = context.Kind.ToString();
-            attributes["key"] = context.Key;
-            attributes["anonymous"] = context.Anonymous;
-
-            return attributes;
-        }
-
-        private static object ValueToObject(LdValue value)
-        {
-            return value.Type switch
-            {
-                LdValueType.Null => null,
-                LdValueType.Bool => value.AsBool,
-                LdValueType.Number => value.AsDouble,
-                LdValueType.String => value.AsString,
-                LdValueType.Array => value.List.Select(ValueToObject).ToList(),
-                LdValueType.Object => value.Dictionary
-                    .Select(kv => new KeyValuePair<string, object>(kv.Key, ValueToObject(kv.Value)))
-                    .ToImmutableDictionary(),
-                _ => null
-            };
+            _client.GetLogger().Warn("Unable to retrieve AI model config for {0}, returning default config", key);
+            return defaultTracker;
         }
 
 
-
-        private const string LdContextVariable = "ldctx";
-
-        /// <summary>
-        /// Retrieves a LaunchDarkly AI config named by the given key.
-        /// </summary>
-        /// <param name="key">the flag key</param>
-        /// <param name="context">the Context</param>
-        /// <param name="defaultValue">the default config, if unable to retrieve</param>
-        /// <param name="variables">the list of variables used when interpolating the prompt</param>
-        /// <returns>an AI config</returns>
-        public LdAiConfigTracker ModelConfig(string key, Context context, LdAiConfig defaultValue,
-            IReadOnlyDictionary<string, object> variables = null)
+        var parsed = ParseConfig(detail.Value, key);
+        if (parsed == null)
         {
+            // ParseConfig already does logging.
+            return defaultTracker;
+        }
 
-            var detail = _client.JsonVariationDetail(key, context, LdValue.Null);
 
-            var defaultTracker = new LdAiConfigTracker(_client, defaultValue, context, key);
-
-            if (detail.IsDefaultValue)
+        var mergedVariables = new Dictionary<string, object> { { LdContextVariable, GetAllAttributes(context) } };
+        if (variables != null)
+        {
+            foreach (var kvp in variables)
             {
-                _client.GetLogger().Warn("No model config available for key {0}", key);
-                return defaultTracker;
-            }
-
-
-            var parsed = ParseConfig(detail.Value, key);
-            if (parsed == null)
-            {
-                // ParseConfig already does logging.
-                return defaultTracker;
-            }
-
-
-            var mergedVariables = new Dictionary<string, object> { { LdContextVariable, GetAllAttributes(context) } };
-            if (variables != null)
-            {
-                foreach (var kvp in variables)
+                if (kvp.Key == LdContextVariable)
                 {
-                    if (kvp.Key == LdContextVariable)
-                    {
-                        _client.GetLogger().Warn("Model config variables contains 'ldctx' key, which is reserved; this key will be the value of the LaunchDarkly context");
-                        continue;
-                    }
-                    mergedVariables[kvp.Key] = kvp.Value;
+                    _client.GetLogger().Warn("AI model config variables contains 'ldctx' key, which is reserved; this key will be the value of the LaunchDarkly context");
+                    continue;
                 }
+                mergedVariables[kvp.Key] = kvp.Value;
             }
-
-
-            var prompt =
-                parsed.Prompt?.Select(m => new LdAiConfig.Message(InterpolateTemplate(m.Content, mergedVariables), m.Role));
-
-            return new LdAiConfigTracker(_client, new LdAiConfig(parsed.Meta?.Enabled ?? false, prompt, parsed.Meta, parsed.Model), context, key);
         }
 
 
+        var prompt =
+            parsed.Prompt?.Select(m => new LdAiConfig.Message(InterpolateTemplate(m.Content, mergedVariables), m.Role));
 
-        private AiConfig ParseConfig(LdValue value, string key)
+        return new LdAiConfigTracker(_client, new LdAiConfig(parsed.Meta?.Enabled ?? false, prompt, parsed.Meta, parsed.Model), context, key);
+    }
+
+    /// <summary>
+    /// Retrieves all attributes from the given context, including private attributes. The attributes
+    /// are converted into C# primitives recursively.
+    /// </summary>
+    /// <param name="context">the context</param>
+    /// <returns>the attributes</returns>
+    private static IDictionary<string, object> GetAllAttributes(Context context)
+    {
+        var attributes = new Dictionary<string, object>();
+        foreach (var key in context.OptionalAttributeNames)
         {
-
-            var serialized = value.ToJsonString();
-            try
-            {
-                return JsonSerializer.Deserialize<AiConfig>(serialized);
-            }
-            catch (JsonException e)
-            {
-                _client.GetLogger().Error("Unable to parse model config for key {0}: {1}", key, e.Message);
-                return null;
-            }
+            attributes[key] = ValueToObject(context.GetValue(AttributeRef.FromLiteral(key)));
         }
 
-        /// <summary>
-        /// THD
-        /// </summary>
-        public void Dispose()
+        attributes["kind"] = context.Kind.ToString();
+        attributes["key"] = context.Key;
+        attributes["anonymous"] = context.Anonymous;
+
+        return attributes;
+    }
+
+    /// <summary>
+    /// Recursively converts an LdValue into a C# object.
+    /// </summary>
+    /// <param name="value">the LdValue</param>
+    /// <returns>the object</returns>
+    private static object ValueToObject(LdValue value)
+    {
+        return value.Type switch
         {
-            _client?.Dispose();
+            LdValueType.Null => null,
+            LdValueType.Bool => value.AsBool,
+            LdValueType.Number => value.AsDouble,
+            LdValueType.String => value.AsString,
+            LdValueType.Array => value.List.Select(ValueToObject).ToList(),
+            LdValueType.Object => value.Dictionary
+                .Select(kv => new KeyValuePair<string, object>(kv.Key, ValueToObject(kv.Value)))
+                .ToImmutableDictionary(),
+            _ => null
+        };
+    }
+
+
+    private static string InterpolateTemplate(string template, IReadOnlyDictionary<string, object> variables)
+    {
+        return Template.Compile(template).Render(variables);
+    }
+
+
+    private AiConfig ParseConfig(LdValue value, string key)
+    {
+
+        var serialized = value.ToJsonString();
+        try
+        {
+            return JsonSerializer.Deserialize<AiConfig>(serialized);
         }
+        catch (JsonException e)
+        {
+            _client.GetLogger().Error("Unable to parse AI model config for key {0}: {1}", key, e.Message);
+            return null;
+        }
+    }
+
+    /// <summary>
+    /// THD
+    /// </summary>
+    public void Dispose()
+    {
+        _client?.Dispose();
     }
 }
