+++ pkgs/sdk/server-ai/src/Metrics/Usage.cs
@@ -1,25 +1,24 @@
 namespace LaunchDarkly.Sdk.Server.Ai.Metrics;
 
-
 /// <summary>
-///
+/// Represents statistics returned by a model provider.
 /// </summary>
-/// <param name="LatencyMs"></param>
+/// <param name="LatencyMs">the duration of the request</param>
 public record struct Statistics(int? LatencyMs);
 
 
 /// <summary>
-///
+/// Represents token usage.
 /// </summary>
-/// <param name="Total"></param>
-/// <param name="Input"></param>
-/// <param name="Output"></param>
+/// <param name="Total">the total tokens used</param>
+/// <param name="Input">the tokens sent as input</param>
+/// <param name="Output">the tokens received as output</param>
 public record struct Usage(int? Total, int? Input, int? Output);
 
 
 /// <summary>
-///
+/// Represents information returned by a model provider.
 /// </summary>
-/// <param name="Usage"></param>
-/// <param name="Statistics"></param>
+/// <param name="Usage">the token usage</param>
+/// <param name="Statistics">the statistics relevant to the request</param>
 public record struct ProviderResponse(Usage? Usage, Statistics? Statistics);
